Make ac_type a 64-bit int.
The aircraft type is a bitmask, and so can be quite large. 64-bits
should be enough for anyone.

+++ include/acinfo.h
@@ -55,7 +55,7 @@ XPLMDataRef aircraft_type = NULL;
 XPLMDataRef aircraft_name = NULL;
 
 /* Values to hold the data. */
-int16_t aircraft_type_value = 0;
+int64_t aircraft_type_value = 0;
 size_t acname_bufsize = 255;
 char aircraft_name_value[ACNAME_BUFSIZE];
